JDK-8309566: Migrate away from TagletWriter and TagletWriterImpl
