+++ src/jdk.javadoc/share/classes/jdk/javadoc/internal/doclets/formats/html/taglets/HtmlDeprecatedTaglet.java
@@ -1,0 +1,91 @@
+/*
+ * Copyright (c) 2023, Oracle and/or its affiliates. All rights reserved.
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+ *
+ * This code is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 only, as
+ * published by the Free Software Foundation.  Oracle designates this
+ * particular file as subject to the "Classpath" exception as provided
+ * by Oracle in the LICENSE file that accompanied this code.
+ *
+ * This code is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+ * version 2 for more details (a copy is included in the LICENSE file that
+ * accompanied this code).
+ *
+ * You should have received a copy of the GNU General Public License version
+ * 2 along with this work; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
+ * or visit www.oracle.com if you need additional information or have any
+ * questions.
+ */
+
+package jdk.javadoc.internal.doclets.formats.html.taglets;
+
+import java.util.List;
+
+import javax.lang.model.element.Element;
+
+import com.sun.source.doctree.DeprecatedTree;
+import com.sun.source.doctree.DocTree;
+
+import jdk.javadoc.internal.doclets.formats.html.HtmlConfiguration;
+import jdk.javadoc.internal.doclets.formats.html.markup.ContentBuilder;
+import jdk.javadoc.internal.doclets.formats.html.markup.HtmlStyle;
+import jdk.javadoc.internal.doclets.formats.html.markup.HtmlTree;
+import jdk.javadoc.internal.doclets.toolkit.Content;
+import jdk.javadoc.internal.doclets.toolkit.taglets.DeprecatedTaglet;
+import jdk.javadoc.internal.doclets.toolkit.taglets.TagletWriter;
+import jdk.javadoc.internal.doclets.toolkit.util.CommentHelper;
+
+public class HtmlDeprecatedTaglet extends DeprecatedTaglet {
+    HtmlDeprecatedTaglet(HtmlConfiguration config) {
+        super(config);
+    }
+
+    @Override
+    public Content getAllBlockTagOutput(Element element, TagletWriter writer) {
+        var w = (TagletWriterImpl) writer;
+        var utils = w.getUtils();
+        var htmlWriter = w.getHtmlWriter();
+
+        ContentBuilder result = new ContentBuilder();
+        CommentHelper ch = utils.getCommentHelper(element);
+        List<? extends DeprecatedTree> deprs = utils.getDeprecatedTrees(element);
+        if (utils.isTypeElement(element)) {
+            if (utils.isDeprecated(element)) {
+                result.add(HtmlTree.SPAN(HtmlStyle.deprecatedLabel,
+                        htmlWriter.getDeprecatedPhrase(element)));
+                if (!deprs.isEmpty()) {
+                    List<? extends DocTree> commentTrees = ch.getDescription(deprs.get(0));
+                    if (!commentTrees.isEmpty()) {
+                        result.add(w.commentTagsToOutput(element, null, commentTrees, false));
+                    }
+                }
+            }
+        } else {
+            if (utils.isDeprecated(element)) {
+                result.add(HtmlTree.SPAN(HtmlStyle.deprecatedLabel,
+                        w.getHtmlWriter().getDeprecatedPhrase(element)));
+                if (!deprs.isEmpty()) {
+                    List<? extends DocTree> bodyTrees = ch.getBody(deprs.get(0));
+                    Content body = w.commentTagsToOutput(element, null, bodyTrees, false);
+                    if (!body.isEmpty())
+                        result.add(HtmlTree.DIV(HtmlStyle.deprecationComment, body));
+                }
+            } else {
+                Element ee = utils.getEnclosingTypeElement(element);
+                if (utils.isDeprecated(ee)) {
+                    result.add(HtmlTree.SPAN(HtmlStyle.deprecatedLabel,
+                            htmlWriter.getDeprecatedPhrase(ee)));
+                }
+            }
+        }
+        return result;
+
+    }
+
+}
