+++ src/jdk.javadoc/share/classes/jdk/javadoc/internal/doclets/formats/html/taglets/HtmlIndexTaglet.java
@@ -1,0 +1,68 @@
+/*
+ * Copyright (c) 2023, Oracle and/or its affiliates. All rights reserved.
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+ *
+ * This code is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 only, as
+ * published by the Free Software Foundation.  Oracle designates this
+ * particular file as subject to the "Classpath" exception as provided
+ * by Oracle in the LICENSE file that accompanied this code.
+ *
+ * This code is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+ * version 2 for more details (a copy is included in the LICENSE file that
+ * accompanied this code).
+ *
+ * You should have received a copy of the GNU General Public License version
+ * 2 along with this work; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
+ * or visit www.oracle.com if you need additional information or have any
+ * questions.
+ */
+
+package jdk.javadoc.internal.doclets.formats.html.taglets;
+
+import javax.lang.model.element.Element;
+
+import com.sun.source.doctree.DocTree;
+import com.sun.source.doctree.IndexTree;
+import com.sun.source.doctree.TextTree;
+
+import jdk.javadoc.internal.doclets.formats.html.HtmlConfiguration;
+import jdk.javadoc.internal.doclets.toolkit.Content;
+import jdk.javadoc.internal.doclets.toolkit.taglets.IndexTaglet;
+import jdk.javadoc.internal.doclets.toolkit.taglets.TagletWriter;
+
+public class HtmlIndexTaglet extends IndexTaglet {
+    HtmlIndexTaglet(HtmlConfiguration config) {
+        super(config);
+    }
+
+    @Override
+    public Content getInlineTagOutput(Element element, DocTree tag, TagletWriter writer) {
+        var indexTree = (IndexTree) tag;
+        var w = (TagletWriterImpl) writer;
+
+        DocTree searchTerm = indexTree.getSearchTerm();
+        String tagText = (searchTerm instanceof TextTree tt) ? tt.getBody() : "";
+        if (tagText.charAt(0) == '"' && tagText.charAt(tagText.length() - 1) == '"') {
+            tagText = tagText.substring(1, tagText.length() - 1);
+        }
+        tagText = tagText.replaceAll("\\s+", " ");
+
+        Content desc = w.getHtmlWriter().commentTagsToContent(element, indexTree.getDescription(), w.getContext().within(indexTree));
+        String descText = extractText(desc);
+
+        return w.createAnchorAndSearchIndex(element, tagText, descText, tag);
+    }
+
+
+    // ugly but simple;
+    // alternatives would be to walk the Content's tree structure, or to add new functionality to Content
+    private String extractText(Content c) {
+        return c.toString().replaceAll("<[^>]+>", "");
+    }
+}
