+++ src/jdk.javadoc/share/classes/jdk/javadoc/internal/doclets/toolkit/taglets/DeprecatedTaglet.java
@@ -26,24 +26,25 @@
 package jdk.javadoc.internal.doclets.toolkit.taglets;
 
 import java.util.EnumSet;
+
 import javax.lang.model.element.Element;
 
 import com.sun.source.doctree.DocTree;
+
 import jdk.javadoc.doclet.Taglet.Location;
+import jdk.javadoc.internal.doclets.toolkit.BaseConfiguration;
 import jdk.javadoc.internal.doclets.toolkit.Content;
 
 /**
  * A taglet that represents the {@code @deprecated} tag.
  */
-public class DeprecatedTaglet extends BaseTaglet {
+public abstract class DeprecatedTaglet extends BaseTaglet {
 
-    public DeprecatedTaglet() {
-        super(DocTree.Kind.DEPRECATED, false,
+    protected DeprecatedTaglet(BaseConfiguration config) {
+        super(config, DocTree.Kind.DEPRECATED, false,
                 EnumSet.of(Location.MODULE, Location.TYPE, Location.CONSTRUCTOR, Location.METHOD, Location.FIELD));
     }
 
     @Override
-    public Content getAllBlockTagOutput(Element holder, TagletWriter writer) {
-        return writer.deprecatedTagOutput(holder);
-    }
+    public abstract Content getAllBlockTagOutput(Element holder, TagletWriter writer);
 }
