+++ src/jdk.javadoc/share/classes/jdk/javadoc/internal/doclets/toolkit/taglets/LiteralTaglet.java
@@ -26,11 +26,13 @@
 package jdk.javadoc.internal.doclets.toolkit.taglets;
 
 import java.util.EnumSet;
+
 import javax.lang.model.element.Element;
 
 import com.sun.source.doctree.DocTree;
-import com.sun.source.doctree.LiteralTree;
+
 import jdk.javadoc.doclet.Taglet.Location;
+import jdk.javadoc.internal.doclets.toolkit.BaseConfiguration;
 import jdk.javadoc.internal.doclets.toolkit.Content;
 
 /**
@@ -40,14 +42,12 @@ import jdk.javadoc.internal.doclets.toolkit.Content;
  * displays as:
  * <blockquote>  {@literal a<B>c}  </blockquote>
  */
-public class LiteralTaglet extends BaseTaglet {
+public abstract class LiteralTaglet extends BaseTaglet {
 
-    LiteralTaglet() {
-        super(DocTree.Kind.LITERAL, true, EnumSet.allOf(Location.class));
+    protected LiteralTaglet(BaseConfiguration config) {
+        super(config, DocTree.Kind.LITERAL, true, EnumSet.allOf(Location.class));
     }
 
     @Override
-    public Content getInlineTagOutput(Element e, DocTree tag, TagletWriter writer) {
-        return writer.literalTagOutput(e, (LiteralTree) tag);
-    }
+    public abstract Content getInlineTagOutput(Element e, DocTree tag, TagletWriter writer);
 }
