+++ src/jdk.javadoc/share/classes/jdk/javadoc/internal/doclets/toolkit/taglets/SummaryTaglet.java
@@ -31,6 +31,8 @@ import javax.lang.model.element.Element;
 import com.sun.source.doctree.DocTree;
 import jdk.javadoc.doclet.Taglet.Location;
 import com.sun.source.doctree.SummaryTree;
+
+import jdk.javadoc.internal.doclets.toolkit.BaseConfiguration;
 import jdk.javadoc.internal.doclets.toolkit.Content;
 
 /**
@@ -38,13 +40,13 @@ import jdk.javadoc.internal.doclets.toolkit.Content;
  */
 public class SummaryTaglet extends BaseTaglet {
 
-    public SummaryTaglet() {
-        super(DocTree.Kind.SUMMARY, true, EnumSet.allOf(Location.class));
+    public SummaryTaglet(BaseConfiguration config) {
+        super(config, DocTree.Kind.SUMMARY, true, EnumSet.allOf(Location.class));
     }
 
     @Override
     public Content getInlineTagOutput(Element holder, DocTree tag, TagletWriter writer) {
         return writer.commentTagsToOutput(holder, tag, ((SummaryTree)tag).getSummary(),
-                writer.isFirstSentence);
+                writer.context.isFirstSentence);
     }
 }
